Fix very long startup on Windows XP in some hardware configurations
Theoretically time issues could have happened on newe configurations too.
It was RF bug that was hackfixed by developers in the game. Dash Faction
has some of those hackfixes removed and because of that the bug was triggered.
It happens to exist when QPC frequency has 31 bit set (sign bit for 32 bit
number).
Also FPS limit accuracy is improved because sleep for less than 1ms is
not supported by windows. Also sleep value can be rounded up on some
Windows versions