Remove README page from the setup and enable welcome page
It README looks ugly in the setup because InnoSetup does not understand
MarkDown and displays it using a variable-width font.
It also has a little benefit to show stuff like features because user
installing the software most likely already know about them and is
probably willing to install as quickly as possible.