Fix writing corrupted save files when autosaving some levels
Game marks some objects as deleted if meshes are not available in new level.
If such half transitioned, half deleted object is saved it can cause crashes
after load. For example L7S3 to L7S4 transition with a guard corpse causes
such problem. To make sure everything is pretty stable and all initial
events got processed wait 5 frames before saving.
Also check if we are in a bomb defuse state. Loading a save from this part
of the game is not working properly