d3d11: use far clip plane and refactor projections
RF mainly uses one far clip plane from fog, but there are exceptions:
* skyroom is rendered without far clip plane
* glares and volumetric lights are using doubled far clip plane
Because it is not possible to change projection matrix between objects
(it would break Z-ordering) a different approach was taken. Disable depth
clipping in rasterizer if game wants far clip plane that is farther than
fog clip plane. Frustum culling should deal with glares and volumetric
lights good enough.
Add abstractions for projection to simplify changing it in the future.
Also get rid of zm variable (set it to 1) to simplify the code.